Episode 19: Limiting Frame Time
Interestingly, although for Chris this was a Windows-specific bug, it actually applies to our version as well. Because requestAnimationFrame does not fire while the target browser view is not visible, this has the same effect as SDL suspending processing while the window is being moved. We use exactly the same technique and it totally works. Awesome.